+++ docs/releasenotes/releasenotes.rst
@@ -12,6 +12,7 @@ Enhancements
 ~~~~~~~~~~~~
 
 * Graph: Enable graphs for Lux device type
+* Popup window: Add ``newwindow: 5`` to open an url as image instead of iframe (doc to be updated)
 
 3.7.0 Beta (13-12-2020)
 -----------------------